Add functions for createing and accessing exercise sets

+++ app/src/main/java/com/example/trainingtracker/model/ExerciseSet.kt
@@ -1,7 +1,7 @@
 package com.example.trainingtracker.model
 
 class ExerciseSet(val id : Int = 0) {
-    var weight : Float = 0f;
+    var weight : Double = 0.0;
     var reps : Int = 0;
     var orderIndex = 0;
     var isWarmup : Boolean = false;
